+++ tests/Client/Decorators/RecordResponseDecoratorTest.php
@@ -1,16 +1,17 @@
 <?php declare(strict_types=1);
 
-namespace Somnambulist\ApiClient\Tests\Client;
+namespace Somnambulist\ApiClient\Tests\Client\Decorators;
 
 use IlluminateAgnostic\Str\Support\Str;
 use PHPUnit\Framework\TestCase;
 use RecursiveDirectoryIterator;
 use RecursiveIteratorIterator;
 use RuntimeException;
+use Somnambulist\ApiClient\Client\ApiClient;
 use Somnambulist\ApiClient\Client\ApiRoute;
 use Somnambulist\ApiClient\Client\ApiRouter;
 use Somnambulist\ApiClient\Client\ApiService;
-use Somnambulist\ApiClient\Client\RecordingApiClient;
+use Somnambulist\ApiClient\Client\Decorators\RecordResponseDecorator;
 use Somnambulist\ApiClient\EntityLocator;
 use Somnambulist\ApiClient\Tests\Stubs\Entities\User;
 use Somnambulist\ApiClient\Tests\Support\Behaviours\UseFactory;
@@ -20,16 +21,15 @@ use Symfony\Component\HttpClient\Response\MockResponse;
 use Symfony\Component\Routing\RouteCollection;
 use function dirname;
 use function file_get_contents;
-use function glob;
 use function rmdir;
 
 /**
- * Class RecordingApiClientTest
+ * Class RecordResponseDecoratorTest
  *
- * @package    Somnambulist\ApiClient\Tests\Client
- * @subpackage Somnambulist\ApiClient\Tests\Client\RecordingApiClientTest
+ * @package    Somnambulist\ApiClient\Tests\Client\Decorators
+ * @subpackage Somnambulist\ApiClient\Tests\Client\Decorators\RecordResponseDecoratorTest
  */
-class RecordingApiClientTest extends TestCase
+class RecordResponseDecoratorTest extends TestCase
 {
 
     use UseFactory;
@@ -47,32 +47,21 @@ class RecordingApiClientTest extends TestCase
     protected function setUp(): void
     {
         $host = 'http://api.example.dev/users/v1';
-        $default = new MockResponse(file_get_contents(__DIR__ . '/../Stubs/user_list.json'), [
+
+        $view = new MockResponse(file_get_contents(__DIR__ . '/../../Stubs/user.json'), [
             'http_code'        => 200,
             'response_headers' => [
                 'Content-Type' => 'application/json',
             ],
         ]);
-        $view = new MockResponse(file_get_contents(__DIR__ . '/../Stubs/user.json'), [
-            'http_code'        => 200,
-            'response_headers' => [
-                'Content-Type' => 'application/json',
-            ],
-        ]);
-        $findById = new MockResponse(file_get_contents(__DIR__ . '/../Stubs/user_list_single.json'), [
-            'http_code'        => 200,
-            'response_headers' => [
-                'Content-Type' => 'application/json',
-            ],
-        ]);
-        $noResults = new MockResponse(file_get_contents(__DIR__ . '/../Stubs/user_list_no_result.json'), [
+        $findById = new MockResponse(file_get_contents(__DIR__ . '/../../Stubs/user_list_single.json'), [
             'http_code'        => 200,
             'response_headers' => [
                 'Content-Type' => 'application/json',
             ],
         ]);
 
-        $callback = function ($method, $url, $options) use ($default, $view, $findById, $noResults) {
+        $callback = function ($method, $url, $options) use ($view, $findById) {
             switch ($url) {
                 case Str::contains($url, '/users/c8259b3b-8603-3098-8361-425325078c9a'):
                     return $view;
@@ -82,12 +71,6 @@ class RecordingApiClientTest extends TestCase
                 case Str::contains($url, '/users?id=c8259b3b-8603-3098-8361-425325078c9a&order=-name,created_at'):
                 case Str::contains($url, '/users?id=c8259b3b-8603-3098-8361-425325078c9a&include=addresses,contacts'):
                     return $findById;
-
-                case Str::contains($url, '/users?id=5715229a-c9d3-4dd9-88f7-e6a1a66f5d31&per_page=1&page=1'):
-                    return $noResults;
-
-                default:
-                    return $default;
             }
         };
         $client = new MockHttpClient($callback);
@@ -96,8 +79,8 @@ class RecordingApiClientTest extends TestCase
         $router->routes()->add('users.list', new ApiRoute('/users'));
         $router->routes()->add('users.view', new ApiRoute('/users/{id}', ['id' => '[0-9a-f\-]{36}']));
 
-        $client = new RecordingApiClient($client, $router);
-        $client->setStore($this->store = dirname(__DIR__, 2) . '/var/cache')->record();
+        $client = new RecordResponseDecorator(new ApiClient($client, $router));
+        $client->setStore($this->store = dirname(__DIR__, 3) . '/var/cache')->record();
 
         $this->locator = new EntityLocator($client, $this->factory()->makeUserMapper(), User::class);
     }
@@ -124,8 +107,8 @@ class RecordingApiClientTest extends TestCase
         $this->locator->find($id = 'c8259b3b-8603-3098-8361-425325078c9a');
         $this->locator->findBy(['id' => $id], ['name' => 'DESC', 'created_at' => 'asc']);
 
-        $this->assertFileExists($this->store . '/15/3a/153a46dff068e201e2f93de7725929800f18b749_1.json');
-        $this->assertFileExists($this->store . '/21/1d/211da158d83dee04816a76e62ccb8c697311009e_1.json');
+        $this->assertFileExists($this->store . '/ac/7e/ac7ebb8e9b2316625ce0bd99b7f02b34f58e3f29_1.json');
+        $this->assertFileExists($this->store . '/df/a8/dfa8715f8b64d05831e340cb091461210efca04b_1.json');
     }
 
     public function testMultipleSameRequestsGetSeparateFiles()
@@ -134,9 +117,9 @@ class RecordingApiClientTest extends TestCase
         $this->locator->find($id = 'c8259b3b-8603-3098-8361-425325078c9a');
         $this->locator->find($id = 'c8259b3b-8603-3098-8361-425325078c9a');
 
-        $this->assertFileExists($this->store . '/15/3a/153a46dff068e201e2f93de7725929800f18b749_1.json');
-        $this->assertFileExists($this->store . '/15/3a/153a46dff068e201e2f93de7725929800f18b749_2.json');
-        $this->assertFileExists($this->store . '/15/3a/153a46dff068e201e2f93de7725929800f18b749_3.json');
+        $this->assertFileExists($this->store . '/ac/7e/ac7ebb8e9b2316625ce0bd99b7f02b34f58e3f29_1.json');
+        $this->assertFileExists($this->store . '/ac/7e/ac7ebb8e9b2316625ce0bd99b7f02b34f58e3f29_2.json');
+        $this->assertFileExists($this->store . '/ac/7e/ac7ebb8e9b2316625ce0bd99b7f02b34f58e3f29_3.json');
     }
 
     public function testCanPlayback()
@@ -144,7 +127,7 @@ class RecordingApiClientTest extends TestCase
         $this->locator->find($id = 'c8259b3b-8603-3098-8361-425325078c9a');
 
         // record the request to avoid needing another stub file
-        $this->assertFileExists($this->store . '/15/3a/153a46dff068e201e2f93de7725929800f18b749_1.json');
+        $this->assertFileExists($this->store . '/ac/7e/ac7ebb8e9b2316625ce0bd99b7f02b34f58e3f29_1.json');
 
         // reset the request tracker so the hash is still the first request: _1
         $this->locator->getClient()->playback()->reset();
@@ -158,7 +141,7 @@ class RecordingApiClientTest extends TestCase
     {
         $this->locator->find($id = 'c8259b3b-8603-3098-8361-425325078c9a');
 
-        $this->assertFileExists($this->store . '/15/3a/153a46dff068e201e2f93de7725929800f18b749_1.json');
+        $this->assertFileExists($this->store . '/ac/7e/ac7ebb8e9b2316625ce0bd99b7f02b34f58e3f29_1.json');
 
         $this->locator->getClient()->playback()->reset();